Improve precision for repeated functions
Before this we were accepting a repeat interval as an int.
When multiplfying this to calculate delay times the rounding
error would become multiplied.

+++ Apps/AudioAppTemplate/Source/Repeat.h
@@ -8,21 +8,21 @@ namespace AudioApp
     class Repeat
     {
     public:
-        static void repeatFunc(int interval, int count, const std::function<void()>& call) {
+        static void repeatFunc(float interval, int count, const std::function<void()>& call) {
             // TODO: use a single timer here instead?
             // Make first call without overhead of using callAfterDelay.
             call();
             for (int i = 1; i < count; ++i) {
-                juce::Timer::callAfterDelay((i)*interval, call);
+                juce::Timer::callAfterDelay(int((float)i*interval), call);
             }
         }
 
-        static void repeatFunc(int interval, int count, const std::function<void(int index)>& call) {
+        static void repeatFunc(float interval, int count, const std::function<void(int index)>& call) {
             // TODO: use a single timer here instead?
             // Make first call without overhead of using callAfterDelay.
             call(0);
             for (int i = 1; i < count; ++i) {
-                juce::Timer::callAfterDelay((i)*interval, [call, i]{
+                juce::Timer::callAfterDelay(int((float)i*interval), [call, i]{
                     call(i);
                 });
             }
